+++ lib/commonmarker.rb
@@ -1,6 +1,8 @@
 #!/usr/bin/env ruby
 require 'commonmarker/commonmarker'
 require 'commonmarker/config'
+require 'commonmarker/renderer'
+require 'commonmarker/renderer/html_renderer'
 require 'stringio'
 require 'cgi'
 require 'set'
@@ -12,42 +14,6 @@ NODE_TYPES = [:none, :document, :blockquote, :list, :list_item,
               :linebreak, :code, :inline_html,
               :emph, :strong, :link, :image]
 LIST_TYPES = [:no_list, :bullet_list, :ordered_list]
-
-# module CMark
-
-  # attach_function :cmark_node_new, [:node_type], :node
-  # attach_function :cmark_node_free, [:node], :void
-  # attach_function :cmark_node_unlink, [:node], :void
-  # attach_function :cmark_node_insert_before, [:node, :node], :int
-  # attach_function :cmark_node_insert_after, [:node, :node], :int
-  # attach_function :cmark_node_prepend_child, [:node, :node], :int
-  # attach_function :cmark_node_append_child, [:node, :node], :int
-  # attach_function :cmark_markdown_to_html, [:string, :int], :string
-  # attach_function :cmark_render_html, [:node], :string
-  # attach_function :cmark_parse_document, [:string, :int], :node
-  # attach_function :cmark_node_first_child, [:node], :node
-  # attach_function :cmark_node_last_child, [:node], :node
-  # attach_function :cmark_node_parent, [:node], :node
-  # attach_function :cmark_node_next, [:node], :node
-  # attach_function :cmark_node_previous, [:node], :node
-  # attach_function :cmark_node_get_type, [:node], :node_type
-  # attach_function :cmark_node_get_literal, [:node], :string
-  # attach_function :cmark_node_set_literal, [:node, :string], :int
-  # attach_function :cmark_node_get_url, [:node], :string
-  # attach_function :cmark_node_set_url, [:node, :string], :int
-  # attach_function :cmark_node_get_title, [:node], :string
-  # attach_function :cmark_node_set_title, [:node, :string], :int
-  # attach_function :cmark_node_get_header_level, [:node], :int
-  # attach_function :cmark_node_set_header_level, [:node, :int], :int
-  # attach_function :cmark_node_get_list_type, [:node], :list_type
-  # attach_function :cmark_node_set_list_type, [:node, :list_type], :int
-  # attach_function :cmark_node_get_list_start, [:node], :int
-  # attach_function :cmark_node_set_list_start, [:node, :int], :int
-  # attach_function :cmark_node_get_list_tight, [:node], :bool
-  # attach_function :cmark_node_set_list_tight, [:node, :bool], :int
-  # attach_function :cmark_node_get_fence_info, [:node], :string
-  # attach_function :cmark_node_set_fence_info, [:node, :string], :int
-# end
 
 module CommonMarker
   class NodeError < StandardError
@@ -354,234 +320,6 @@ module CommonMarker
     def free
       CMark.node_unlink(@pointer)
       # CMark.node_free(@pointer)
-    end
-  end
-
-  class Renderer
-    attr_accessor :in_tight, :warnings, :in_plain
-    def initialize
-      @stream = StringIO.new
-      @need_blocksep = false
-      @warnings = Set.new []
-      @in_tight = false
-      @in_plain = false
-    end
-
-    def out(*args)
-      args.each do |arg|
-        if arg == :children
-          @node.each_child do |child|
-            self.out(child)
-          end
-        elsif arg.kind_of?(Array)
-          arg.each do |x|
-            self.render(x)
-          end
-        elsif arg.kind_of?(Node)
-          self.render(arg)
-        else
-          @stream.write(arg)
-        end
-      end
-    end
-
-    def render(node)
-      @node = node
-      if node.type == :document
-        self.document(node)
-        return @stream.string
-      elsif self.in_plain && node.type != :text && node.type != :softbreak
-        node.each_child do |child|
-          render(child)
-        end
-      else
-        begin
-          self.send(node.type, node)
-        rescue NoMethodError => e
-          @warnings.add("WARNING:  " + node.type.to_s + " not implemented.")
-          raise e
-        end
-      end
-    end
-
-    def document(node)
-      self.out(:children)
-    end
-
-    def code_block(node)
-      self.code_block(node)
-    end
-
-    def reference_def(node)
-    end
-
-    def blocksep
-      self.out("\n")
-    end
-
-    def containersep
-      if !self.in_tight
-        self.out("\n")
-      end
-    end
-
-    def block(&blk)
-      if @need_blocksep
-        self.blocksep
-      end
-      blk.call
-      @need_blocksep = true
-    end
-
-    def container(starter, ender, &blk)
-      self.out(starter)
-      self.containersep
-      @need_blocksep = false
-      blk.call
-      self.containersep
-      self.out(ender)
-    end
-
-    def plain(&blk)
-      old_in_plain = @in_plain
-      @in_plain = true
-      blk.call
-      @in_plain = old_in_plain
-    end
-  end
-
-  class HtmlRenderer < Renderer
-    def render(node)
-      result = super(node)
-      if node.type == :document
-        result += "\n"
-      end
-    end
-
-    def header(node)
-      block do
-        self.out("<h", node.header_level, ">", :children,
-               "</h", node.header_level, ">")
-      end
-    end
-
-    def paragraph(node)
-      block do
-        if self.in_tight
-          self.out(:children)
-        else
-          self.out("<p>", :children, "</p>")
-        end
-      end
-    end
-
-    def list(node)
-      old_in_tight = self.in_tight
-      self.in_tight = node.list_tight
-      block do
-        if node.list_type == :bullet_list
-          container("<ul>", "</ul>") do
-            self.out(:children)
-          end
-        else
-          start = node.list_start == 1 ? '' :
-                  (' start="' + node.list_start.to_s + '"')
-          container(start, "</ol>") do
-            self.out(:children)
-          end
-        end
-      end
-      self.in_tight = old_in_tight
-    end
-
-    def list_item(node)
-      block do
-        container("<li>", "</li>") do
-          self.out(:children)
-        end
-      end
-    end
-
-    def blockquote(node)
-      block do
-        container("<blockquote>", "</blockquote>") do
-          self.out(:children)
-        end
-      end
-    end
-
-    def hrule(node)
-      block do
-        self.out("<hr />")
-      end
-    end
-
-    def code_block(node)
-      block do
-        self.out("<pre><code")
-        if node.fence_info && node.fence_info.length > 0
-          self.out(" class=\"language-", node.fence_info.split(/\s+/)[0], "\">")
-        else
-          self.out(">")
-        end
-        self.out(CGI.escapeHTML(node.string_content))
-        self.out("</code></pre>")
-      end
-    end
-
-    def html(node)
-      block do
-        self.out(node.string_content)
-      end
-    end
-
-    def inline_html(node)
-      self.out(node.string_content)
-    end
-
-    def emph(node)
-      self.out("<em>", :children, "</em>")
-    end
-
-    def strong(node)
-      self.out("<strong>", :children, "</strong>")
-    end
-
-    def link(node)
-      self.out('<a href="', node.url.nil? ? '' : URI.escape(node.url), '"')
-      if node.title && node.title.length > 0
-        self.out(' title="', CGI.escapeHTML(node.title), '"')
-      end
-      self.out('>', :children, '</a>')
-    end
-
-    def image(node)
-      self.out('<img src="', URI.escape(node.url), '"')
-      if node.title && node.title.length > 0
-        self.out(' title="', CGI.escapeHTML(node.title), '"')
-      end
-      plain do
-        self.out(' alt="', :children, '" />')
-      end
-    end
-
-    def text(node)
-      self.out(CGI.escapeHTML(node.string_content))
-    end
-
-    def code(node)
-      self.out("<code>")
-      self.out(CGI.escapeHTML(node.string_content))
-      self.out("</code>")
-    end
-
-    def linebreak(node)
-      self.out("<br/>")
-      self.softbreak(node)
-    end
-
-    def softbreak(node)
-      self.out("\n")
     end
   end
